第 40 轮回归成果：权威 md 733→705（累计 3324→705；monitor -17、dbmw -4、guild -4、statics -2）；回滚 2 处 R40 误改（VillageAttackedManager tGMConfig 改名、SendToClient register 形态）

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/DNFTowerRank.cpp
@@ -103,8 +103,8 @@ unsigned int CTowerRank::getRankData(unsigned int floor, const char* name, unsig
 
 void CTowerRank::reset()
 {
-    unsigned int i = 0;
-    while (i <= 3)
+    int i = 0;
+    while ((unsigned int)i <= 3)
     {
         m_ranks[i].clear();
         i++;
